backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# QR Code generation
qrcode[pil]==7.4.2

# HTTP client (for Telegram validation)
httpx==0.26.0

# Utils
python-dotenv==1.0.1
